+++ packages/ui-components/src/combo-box/combo-box.tsx
@@ -127,6 +127,8 @@ export const Input = styled.input`
 `;
 
 export interface ComboBoxProps extends InteractiveComponentProps<Item> {
+    /** An optional id for the component */
+    id?: string;
     /** Whether to open the select dropdown on load or not, defaults to false */
     initialIsOpen?: boolean;
     /** The items to pick from the list. Each should have a label and a value */
@@ -261,6 +263,7 @@ function ComboBox(props: ComboBoxProps): JSX.Element {
     const inputProps = {
         ...getInputProps({
             disabled: props.disabled,
+            id: props.id,
         }),
         ...getReferenceProps(),
     };